reformatting and breaking down components into seperate files to clean things up further
